More functionalities to gfmWindow; Code can compile; Makefile is a lot safer

+++ include/GFraMe/core/gfmTimer_bkend.h
@@ -11,7 +11,7 @@
 /** "Export" the timer structure's type */
 typedef struct stGFMTimer gfmTimer;
 /** 'Exportable' size of gfmTimer */
-extern const size_t sizeofGFMTimer;
+extern const int sizeofGFMTimer;
 
 /**
  * Get how long each frame must take for the timer function;
